Made BasePoints, Towers and Rounds.
It's kind of working, but there are two major issues. First, I don't
know how to detach sprites from the scene properly. It has to be done on
the update thread, otherwise an out-of-bounds error occurs. Second, due
to multi-threading, there are some synchronization issues that make the
game crash. This sometimes happens, for example, when a round is moving
towards an enemy that was just killed by another round and cannot refer
to its target, since it was removed from the currentEnemies array.